Setup now installs dependencies
While distutils package may be a great solution because setuptools is not built in, this setup method can't understand install_requires, extras_require and python_requires (and entry_points, etc... but it is not in the setup). Importing setuptools fixes these without having to use this module directly. It can be better to force the user to have setuptools but to keep the old method, I've added a try/except condition.